Tailwind v4 + Refactor code for modularity
