+++ src/app/layout.tsx
@@ -2,8 +2,6 @@ import type { Metadata } from "next"
 
 import "./globals.css"
 
-import Navbar from "@/components/navbar"
-
 export const metadata: Metadata = {
   title: "Car Lander",
   description: "Land your car faster",
@@ -16,10 +14,7 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="en">
-      <body>
-        <Navbar />
-        {children}
-      </body>
+      <body>{children}</body>
     </html>
   )
 }
