+++ src/app/login/page.tsx
@@ -4,8 +4,10 @@ import { auth } from "@/lib/auth"
 
 import FirstLogin from "@/components/first-login"
 import LoginForm from "@/components/login-form"
+import OAuthButtons from "@/components/oauth-buttons"
 
 import { Avatar, AvatarFallback, AvatarImage } from "@/ui/avatar"
+import { Separator } from "@/ui/separator"
 
 const itim = Itim({
   weight: "400",
@@ -18,35 +20,40 @@ export default async function LoginPage() {
   const session = await auth()
 
   return (
-    <div className={`${jost.className} flex flex-col items-center`}>
-      <div className="w-full max-w-sm space-y-4">
-        {/* Logo and Title */}
-        <div className="mb-10 flex flex-col items-center">
-          <Avatar className="mt-[4rem] aspect-square h-auto w-max rounded-lg border-4 border-zinc-950">
-            <AvatarImage src="" />
-            <AvatarFallback
-              className={`${itim.className} rounded-none p-2 indent-[.2em] text-xl tracking-[.2em]`}
-            >
-              LOGO
-            </AvatarFallback>
-          </Avatar>
-        </div>
-
-        {/* Welcome Text */}
-        <h1 className="mb-6 mt-10 text-center indent-[.3em] text-2xl uppercase tracking-[.3em]">
+    <div
+      className={`m-auto flex w-max max-w-sm flex-col items-center ${jost.className}`}
+    >
+      <div className="flex flex-col items-center">
+        <Avatar className="mt-[4rem] aspect-square h-auto w-max rounded-lg border-4 border-zinc-950">
+          <AvatarImage src="" />
+          <AvatarFallback
+            className={`${itim.className} rounded-none p-2 indent-[.2em] text-xl tracking-[.2em]`}
+          >
+            LOGO
+          </AvatarFallback>
+        </Avatar>
+        <h1 className="mt-10 mb-6 indent-[.3em] text-2xl tracking-[.3em] uppercase">
           Carlander
         </h1>
-        <div className="text-center text-sm">
-          <span className="text-lime-500">Welcome!</span>
-          <span className="text-lime-500"> Please sign in to continue</span>
-          <span>
-            <br />
-            Sign in with
-          </span>
-        </div>
+        <p className="my-6 text-center">
+          Welcome! Please log in to continue
+          <br />
+          <span className="text-lime-500">Log in using</span>
+        </p>
 
-        {session?.newUser ? <FirstLogin /> : <LoginForm />}
+        {!session?.newUser && <OAuthButtons />}
       </div>
+      {!session?.newUser && (
+        <>
+          <div className="flex w-xs items-center gap-6">
+            <Separator className="shrink border-2 border-lime-500" />
+            or
+            <Separator className="shrink border-2 border-lime-500" />
+          </div>
+        </>
+      )}
+
+      {session?.newUser ? <FirstLogin /> : <LoginForm />}
     </div>
   )
 }
